simulados-v2/backend/requirements.txt
# ── Core ─────────────────────────────────────────────
django==5.0.6
django-ninja==1.1.0
psycopg2-binary==2.9.9
python-decouple==3.8

# ── Auth ──────────────────────────────────────────────
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
bcrypt==4.1.3

# ── Cache / Tasks ─────────────────────────────────────
redis==5.0.4
celery==5.4.0
django-redis==5.4.0

# ── Utils ─────────────────────────────────────────────
Pillow==10.3.0
python-slugify==8.0.4
pydantic==2.7.1

# ── Server ────────────────────────────────────────────
gunicorn==22.0.0
whitenoise==6.6.0

# ── Dev / Test ────────────────────────────────────────
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0
faker==25.0.0
flake8==7.0.0
